fix windows unit tests
